mergeMantaDelly.py - Changed conditions for filtering --filterOverlap
Functions added for conditions for comparison. A seperate one for the insertions (with insertionlength) and one for all other SVs (with endpos and CIEND).
Filtering changed in def compareFilterAndWriteSVs. List similar_SVs is sorted with Manta and the smallest CIs at the top. If the first one in the list is a Manta SV and the second one is one of DELLY, a CSA of 2 and the INFO field of the DELLY call are added to the INFO field of the Manta-SV record.