mergeMantaDelly.py --filterOverlap changed filter for TRAs and BNDs
If a translocation or break endpoint is put in the merged file, it will only be put in once, not for both chromosomes that were involved in the translocation event.
This will be helpful when counting how many SVs there are in total. Note that the first BND of the pair in the list is kept, and the list is sorted on chrom number (21 after 2!).
This is working now for the BNDs, changes should be made to filter for TRAs.