mergeMantaDelly.py -changed def filterBndAndWriteSVs
Since only BNDs are printed for both chromosomes involved in translocation from the Manta output, and the TRA from Delly output are already in the file only once, there does not have to be filtered for TRA in this function.
The lines already in the function to filter for TRA have been removed and the function is renamed.