mergeMantaDelly.py -filterBndAndWriteSVs --changed calculation of overlap Manta and Delly for BNDs
For BNDs, there is filtered to keep only one BND in the list for the event instead of two for both chromosomes in the TRA event (see previous commits).
When this filtering is done, there is now also taken into account that for one chrom there might be a call for Delly as well as Manta (i.e. CSA=2).
Because both chroms are considered to be the same (call of TRA event),
if one is called by 2 tools and the other is not they will be combined and there is stated that this SV is called by both Delly and Manta (thus CSA=2).
This is instead of keeping CSA=1 if the chrom of the event that is shown in the vcf had CSA=1 before combining the two.
This is an important point when wanting to filter the overlapping SVs of Manta and Delly.