mergeMantaDelly.py --def conditionsForComparison -improving conditionsTRA/BND - more improvements may be needed.
To improve the way the translocations and bnds are put in lists, the conditions have to be more specific. There were a few wrong comparisons in the if statements that are now changed.
These conditions might need more improvements, the data for testing if it works does not seem to contain two BNDs or TRAs so the if statement is now only true till a certain point,
the final outcome is always false. More improvements may be found to be necessary.